added scoring on top

+++ projects/lpc40xx_freertos/l4_io/led_matrix/shapes.h
@@ -28,4 +28,4 @@ void draw_friend(int x, int y);
 /**
  * Print score
  */
-void print_score(uint8_t score, int x, int y, led_matrix__color_e shape_color);
+void print_score(uint8_t score, uint8_t x, uint8_t y, led_matrix__color_e shape_color);
